old-tests/webapps/DOMEvents/tests/submissions/TestTWF_Paris/DavidBruant/throwing-in-listener-when-all-have-not-run-yet.html

<!DOCTYPE html>
<!-- Submitted from TestTWF Paris -->
<meta charset="UTF-8">
<title>Throwing in event listener</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<button type="button" id="test">Start test</button>

<script>
    var t = async_test("Throwing in event listener");

    document.addEventListener('DOMContentLoaded', function(){
        var element = document.getElementById('bim');

        var secondCalled = false;

        element.addEventListener('click', function(){
            throw new Error('Error from listener');
        });
        element.addEventListener('click', t.step_func(function(){ secondCalled = true; }), false);

        element.click();

        assert_true(secondCalled);
        t.done();
    });
</script>

<div id="bim">
</div>
